fix: 마이페이지- 내가 작성한 게시글 레이아웃을 커뮤니티 레이아웃으로 적용 #9
